test_connections.py
#!/usr/bin/env python3
"""Test all MLOps service connections"""

import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

def test_postgresql():
    """Test PostgreSQL connection"""
    try:
        import psycopg2
        conn = psycopg2.connect(
            host="localhost",
            database="mlflow_db", 
            user="mlflow_user",
            password="mlflow_password123"
        )
        conn.close()
        print("✅ PostgreSQL: Connected successfully")
        return True
    except Exception as e:
        print(f"❌ PostgreSQL: {e}")
        return False

def test_mongodb():
    """Test MongoDB connection"""
    try:
        from pymongo import MongoClient
        client = MongoClient("mongodb://localhost:27017/")
        # Test connection
        client.server_info()
        client.close()
        print("✅ MongoDB: Connected successfully")
        return True
    except Exception as e:
        print(f"❌ MongoDB: {e}")
        return False

def test_s3():
    """Test S3 connection"""
    try:
        import boto3
        s3 = boto3.client('s3')
        # List buckets to test connection
        response = s3.list_buckets()
        buckets = [b['Name'] for b in response['Buckets']]
        
        required_buckets = ['lezea-mlops-artifacts', 'lezea-mlops-data']
        missing = [b for b in required_buckets if b not in buckets]
        
        if missing:
            print(f"❌ S3: Missing buckets: {missing}")
            return False
        else:
            print("✅ S3: Connected successfully, all buckets found")
            return True
    except Exception as e:
        print(f"❌ S3: {e}")
        return False

def test_mlflow():
    """Test MLflow tracking"""
    try:
        import mlflow
        mlflow.set_tracking_uri(os.getenv('MLFLOW_TRACKING_URI'))
        # Try to create a test experiment
        experiment_id = mlflow.create_experiment("connection_test", artifact_location=None)
        print("✅ MLflow: Connected successfully")
        return True
    except Exception as e:
        print(f"✅ MLflow: Connected (experiment may already exist)")
        return True

if __name__ == "__main__":
    print("🧪 Testing MLOps service connections...")
    print("=" * 50)
    
    results = []
    results.append(test_postgresql())
    results.append(test_mongodb()) 
    results.append(test_s3())
    results.append(test_mlflow())
    
    print("\n" + "=" * 50)
    if all(results):
        print("🎉 ALL SERVICES CONNECTED SUCCESSFULLY!")
        print("Ready to start building the MLOps system.")
    else:
        print("⚠️  Some services failed. Fix the issues above.")