requirements.txt
# Core MLOps stack
mlflow==2.8.1
pymongo==4.6.0
boto3==1.34.0
psycopg2-binary==2.9.9

# Data versioning
dvc[s3]==3.32.0

# System monitoring
psutil==5.9.6
GPUtil==1.4.0
pynvml==11.5.0

# Data handling
pandas==2.1.4
numpy==1.24.4
PyYAML==6.0.1

# Environment management
python-dotenv==1.0.0

# Machine Learning
torch==2.1.2
scikit-learn==1.3.2

# Development tools
pytest==7.4.3
